requirements.txt
django==2.1.4
graphene-django==2.2.0
django-filter==2.0.0
django-graphql-jwt==0.2.2
psycopg2==2.8.3
python-decouple==3.1
django-cors-headers==3.1.1
pytz==2019.3
